requirements.txt
Flask==1.1.2
Flask-JWT-Extended==4.1.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
marshmallow-sqlalchemy==0.24.2
nose==1.3.7
passlib==1.7.4
unittest2==1.1.0
